Add throttle config and BUILD.
